refactor(search): extract SearchPlan and package-visible RRF fusion
Move Reciprocal Rank Fusion onto HybridSearch as package API and compute
lane/window decisions in SearchPlan.make so Wax.search no longer inlines
those pre-I/O choices. vectorOnly still throws at the search shell.
